fixed debt getAll method array error

+++ backend/cs-storage-core/app/Models/Debt.php
@@ -12,7 +12,7 @@ class Debt extends Model
     ];
 
     protected $casts = [
-        'value' => 'decimal'
+        'value' => 'float'
     ];
 
     public function customer(){
